backend/requirements.txt
Django==5.2.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
Pillow==10.4.0
reportlab==4.2.2
python-decouple==3.8
whitenoise==6.8.2
gunicorn==21.2.0
# psycopg2-binary==2.9.9  # For PostgreSQL in production
